requirements.txt
PyPDF2
python-docx
openpyxl
click
pytest
